refactor: useUserStore 통합 (#69)
* refactor: entities/hooks/useUserStore로 유저 상태 관리 훅 통합

useUserStore 훅을 import하고 있는 파일 수정

* refactor: UserAvatar -> UserIcon 컴포넌트로 통합

---------
