+++ src/entities/user/model/hooks/useUserStore.ts
@@ -1,26 +1,14 @@
 import { create } from 'zustand';
 import { createJSONStorage, persist } from 'zustand/middleware';
-
-interface AuthState {
-  user: {
-    profileImgUrl: string;
-    nickname: string;
-  };
-  isLoggedIn: boolean;
-  logIn: (user: { profileImgUrl: string; nickname: string }) => void;
-  logOut: () => void;
-}
+import { User, UserState } from '../types';
 
 const useUserStore = create(
-  persist<AuthState>(
+  persist<UserState>(
     (set) => ({
-      user: {
-        profileImgUrl: '',
-        nickname: '',
-      },
+      user: null,
       isLoggedIn: false,
-      logIn: (user) => set({ isLoggedIn: true, user: user as AuthState['user'] }),
-      logOut: () => set({ isLoggedIn: false, user: undefined }),
+      logIn: (user) => set({ isLoggedIn: true, user: user as User }),
+      logOut: () => set({ isLoggedIn: false, user: null }),
     }),
     {
       // Zustand의 persist 미들웨어를 사용하여 상태를 로컬 스토리지에 저장
