--- src/entities/user/model/user.ts
@@ -1,6 +1,0 @@
-export interface User {
-  email: string;
-  nickname: string;
-  profileImg: string;
-  interest: string[];
-}
