+++ src/entities/user/ui/UserIcon.tsx
@@ -1,18 +1,14 @@
-import { Button, useUserStore } from '@/shared/';
-import { useShallow } from 'zustand/shallow';
+import { Button } from '@/shared';
 import Image from 'next/image';
+import useUserStore from '../model/hooks/useUserStore';
 
-function UserIcon() {
-  const { profileImgUrl } = useUserStore(
-    useShallow((state) => ({
-      profileImgUrl: state.user.profileImgUrl,
-    })),
-  );
+export default function UserIcon() {
+  const user = useUserStore((state) => state.user);
+  if (!user) return null;
+
   return (
     <Button variant="secondary" size="icon" className="relative overflow-hidden rounded-full">
-      <Image src={profileImgUrl} alt="profile-image" fill />
+      <Image src={user.profileImgUrl} alt={`${user.nickname}의 프로필`} fill />
     </Button>
   );
 }
-
-export default UserIcon;
