--- src/entities/user/ui/index.ts
@@ -1,3 +1,0 @@
-import UserIcon from './UserIcon';
-
-export { UserIcon };
