+++ src/features/login-with-kakao/model/useLoginWithKakaoMutation.ts
@@ -1,15 +1,15 @@
 import { useMutation } from '@tanstack/react-query';
-import { useUserStore } from '@/entities/user/model/userStore';
+import { useUserStore } from '@/entities/user';
 import { loginWithKakao } from '../api/loginWithKakao';
 
 export function useLoginWithKakaoMutation() {
-  const setUser = useUserStore((state) => state.setUser);
+  const { logIn } = useUserStore((state) => state);
 
   return useMutation({
     mutationFn: loginWithKakao,
     onSuccess: (data) => {
       if (!data) return;
-      setUser(data);
+      logIn(data);
     },
   });
 }
