+++ src/shared/index.ts
@@ -1,3 +1,2 @@
 export * from './ui';
-export * from './lib/hooks';
 export * from './api';
