--- src/shared/lib/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useUserStore } from './useUserStore';
